feat(test-numbers): only use numbers for tests without titles (#19)
This was super annoying. Any time I added/moved/removed a test, all the
snapshots would need updating because the title for every one of them
changed. Pretty annoying. So now we only use the number for untitled
tests and we only increment it for the untitled tests.

BREAKING CHANGE: Your snapshots will break (for the better). Just update them. Nothing else changed.