fix: Allow formating error in snapshot
For fixtures expected to fail, an error in the snapshot contains a full path. This is a problem with CI because the paths are different. It could be normalized with `formatResult`, but that function was not called for errors. This PR fixes that and calls `formatResult` with the message of an error.